refactor(events): split organization_service god module into a package (#578)
organization_service.py was a ~759 LOC, 27-function god service mixing
membership, tokens/invitations, billing/settings, and contact messages.
Decompose it into a package of concern-focused submodules:

- lifecycle.py  — create/update organization
- membership.py — membership requests, members, staff
- tokens.py     — organization tokens + invitation claiming
- contact.py    — contact email verification, method, and messages

Every public symbol is re-exported from __init__.py so the historical
`from events.service import organization_service` import path (and
`organization_service.<name>` access) keeps working — no call sites change.
Pure refactor, no behavior change. mypy --strict clean.
