--- src/events/service/organization_service.py
@@ -1,759 +1,0 @@
-import typing as t
-from datetime import timedelta
-from uuid import UUID
-
-from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.db import transaction
-from django.db.models import F, Q
-from django.shortcuts import get_object_or_404
-from django.utils import timezone
-from django.utils.translation import gettext_lazy as _
-from ninja.errors import HttpError
-
-from accounts.jwt import blacklist as blacklist_token
-from accounts.jwt import check_blacklist, create_token
-from accounts.models import RevelUser
-from accounts.service.account import token_to_payload
-from common.models import SiteSettings
-from events import models, schema, tasks
-from events.exceptions import (
-    AlreadyMemberError,
-    OrganizationTokenGrantInvariantError,
-    OrganizationTokenMembershipTierRequiredError,
-    OrganizationTokenStaffGrantForbidden,
-    PendingMembershipRequestExistsError,
-    RevenueReportCadenceOwnerOnlyError,
-)
-from events.models import (
-    MembershipTier,
-    Organization,
-    OrganizationContactMessage,
-    OrganizationMember,
-    OrganizationMembershipRequest,
-    OrganizationStaff,
-    OrganizationToken,
-    PermissionsSchema,
-)
-from events.models.organization import _get_default_permissions
-from events.service import blacklist_service
-from events.utils.reserved_slug_tokens import find_reserved_token
-from notifications.enums import NotificationType
-from notifications.signals import notification_requested
-
-# Canned messages for the organization-token guard exceptions. Co-located with
-# the service that raises them so the per-app exception handlers (and any caller)
-# render an identical, translatable ``{"detail": ...}`` body.
-STAFF_GRANT_FORBIDDEN_MESSAGE = _("Only the organization owner can manage staff-granting tokens.")
-GRANT_INVARIANT_MESSAGE = _("At least one of grants_membership or grants_staff_status must be True.")
-MEMBERSHIP_TIER_REQUIRED_MESSAGE = _("membership_tier_id is required when grants_membership is True.")
-REVENUE_CADENCE_OWNER_ONLY_MESSAGE = _("Only the organization owner can change the revenue report cadence.")
-
-
-def _create_and_send_contact_email_verification(
-    organization: Organization,
-    email: str,
-    user: RevelUser,
-) -> str:
-    """Create verification token and send verification email.
-
-    Args:
-        organization: The organization
-        email: The email to verify
-        user: The user associated with the verification
-
-    Returns:
-        The verification token
-    """
-    verification_payload = schema.VerifyOrganizationContactEmailJWTPayloadSchema(
-        organization_id=organization.id,
-        user_id=user.id,
-        email=email,
-        exp=timezone.now() + settings.VERIFY_TOKEN_LIFETIME,
-    )
-    token = create_token(verification_payload.model_dump(mode="json"), settings.SECRET_KEY, settings.JWT_ALGORITHM)
-
-    # Send verification email
-    def send_verification_email() -> None:
-        tasks.send_organization_contact_email_verification.delay(
-            email=email,
-            token=token,
-            organization_name=organization.name,
-            organization_slug=organization.slug,
-        )
-
-    transaction.on_commit(send_verification_email)
-    return token
-
-
-@transaction.atomic
-def create_organization(
-    owner: RevelUser,
-    name: str,
-    contact_email: str,
-    description: str | None = None,
-    city_id: int | None = None,
-    address: str | None = None,
-    location_maps_url: str | None = None,
-    location_maps_embed: str | None = None,
-    slug: str | None = None,
-) -> Organization:
-    """Create a new organization.
-
-    Args:
-        owner: The user who will own the organization
-        name: The name of the organization
-        contact_email: The contact email for the organization
-        description: Optional description for the organization
-        city_id: Optional city id for the organization
-        address: Optional address for the organization
-        location_maps_url: Optional shareable Google Maps URL
-        location_maps_embed: Optional Google Maps embed URL for iframe
-        slug: Optional explicit slug; when omitted the model derives one from the name.
-
-    Returns:
-        The created Organization instance
-
-    Raises:
-        HttpError: If the user already owns an organization
-    """
-    if offending := find_reserved_token(name):
-        raise HttpError(
-            400,
-            str(_('The name contains a reserved word: "%(token)s". Please choose a different name.'))
-            % {"token": offending},
-        )
-    # Check if user already owns an organization
-    if Organization.objects.filter(owner=owner).exists():
-        raise HttpError(
-            400, str(_("You already own an organization. Only one organization per user as owner is allowed."))
-        )
-
-    # Check if contact email matches user email and is verified
-    contact_email_verified = False
-    if contact_email.lower() == owner.email.lower() and owner.email_verified:
-        contact_email_verified = True
-
-    # Create the organization
-    organization = Organization.objects.create(
-        name=name,
-        slug=slug or "",  # empty → the SluggedModel mixin derives it from the name
-        owner=owner,
-        description=description or "",
-        contact_email=contact_email,
-        contact_email_verified=contact_email_verified,
-        visibility=Organization.Visibility.STAFF_ONLY,  # Default to staff only
-        city_id=city_id,
-        address=address,
-        location_maps_url=location_maps_url,
-        location_maps_embed=location_maps_embed,
-    )
-
-    # Send verification email if contact email is not auto-verified
-    if not contact_email_verified:
-        _create_and_send_contact_email_verification(organization, contact_email, owner)
-
-    return organization
-
-
-def create_membership_request(
-    organization: Organization, user: RevelUser, message: str | None = None
-) -> OrganizationMembershipRequest:
-    """Create a membership request.
-
-    Args:
-        organization: The organization to request membership for.
-        user: The user requesting membership.
-        message: Optional message from the user.
-
-    Returns:
-        The created OrganizationMembershipRequest instance.
-
-    Raises:
-        HttpError: If the organization does not accept requests, or the user is blacklisted.
-        AlreadyMemberError: If the user is already a member.
-        PendingMembershipRequestExistsError: If a pending request already exists.
-    """
-    if not organization.accept_membership_requests:
-        raise HttpError(400, str(_("The organization does not accept new members.")))
-
-    if blacklist_service.check_user_hard_blacklisted(user, organization):
-        raise HttpError(403, str(_("You are not allowed to request membership for this organization.")))
-
-    if models.OrganizationMember.objects.filter(organization=organization, user=user).exists():
-        raise AlreadyMemberError
-
-    if OrganizationMembershipRequest.objects.filter(
-        organization=organization, user=user, status=OrganizationMembershipRequest.Status.PENDING
-    ).exists():
-        raise PendingMembershipRequestExistsError
-
-    return OrganizationMembershipRequest.objects.create(organization=organization, user=user, message=message)
-
-
-@transaction.atomic
-def approve_membership_request(
-    membership_request: models.OrganizationMembershipRequest, decided_by: RevelUser, tier: MembershipTier
-) -> None:
-    """Approve a membership request and assign tier.
-
-    Args:
-        membership_request: The membership request to approve
-        decided_by: The user approving the request
-        tier: The membership tier to assign
-    """
-    membership_request.status = models.OrganizationMembershipRequest.Status.APPROVED
-    membership_request.decided_by = decided_by
-    membership_request.save(update_fields=["status", "decided_by"])
-
-    # Create or update membership with tier (this will trigger MEMBERSHIP_GRANTED notification via signal)
-    # Use update_or_create to ensure clean() method is called for validation
-    member, created = models.OrganizationMember.objects.update_or_create(
-        organization=membership_request.organization,
-        user=membership_request.user,
-        defaults={"tier": tier, "status": OrganizationMember.MembershipStatus.ACTIVE},
-    )
-
-    # Explicitly call clean to validate tier belongs to same organization
-    member.full_clean()
-
-    # Send approval notification
-    def send_approval_notification() -> None:
-        frontend_base_url = SiteSettings.get_solo().frontend_base_url
-        notification_requested.send(
-            sender=models.OrganizationMembershipRequest,
-            user=membership_request.user,
-            notification_type=NotificationType.MEMBERSHIP_REQUEST_APPROVED,
-            context={
-                "organization_id": str(membership_request.organization_id),
-                "organization_name": membership_request.organization.name,
-                "frontend_url": f"{frontend_base_url}/org/{membership_request.organization.slug}",
-            },
-        )
-
-    if created:
-        transaction.on_commit(send_approval_notification)
-
-
-def reject_membership_request(request: models.OrganizationMembershipRequest, decided_by: RevelUser) -> None:
-    """Reject a membership request."""
-    request.status = models.OrganizationMembershipRequest.Status.REJECTED
-    request.decided_by = decided_by
-    request.save(update_fields=["status", "decided_by"])
-
-    # Send rejection notification
-    def send_rejection_notification() -> None:
-        frontend_base_url = SiteSettings.get_solo().frontend_base_url
-        notification_requested.send(
-            sender=models.OrganizationMembershipRequest,
-            user=request.user,
-            notification_type=NotificationType.MEMBERSHIP_REQUEST_REJECTED,
-            context={
-                "organization_id": str(request.organization_id),
-                "organization_name": request.organization.name,
-                "frontend_url": f"{frontend_base_url}/organizations",
-            },
-        )
-
-    transaction.on_commit(send_rejection_notification)
-
-
-def create_organization_token(
-    *,
-    organization: Organization,
-    issuer: RevelUser,
-    duration: timedelta | int = 60,
-    grants_membership: bool = True,
-    membership_tier: MembershipTier | None = None,
-    grants_staff_status: bool = False,
-    name: str | None = None,
-    max_uses: int = 0,
-) -> OrganizationToken:
-    """Create an organization token for sharing invitation links.
-
-    Args:
-        organization: The organization the token belongs to.
-        issuer: The user creating the token.
-        duration: Token validity duration in minutes (or timedelta).
-        grants_membership: Whether the token grants membership.
-        membership_tier: The membership tier to assign (required if grants_membership is True).
-        grants_staff_status: Whether the token grants staff status.
-        name: Display name for the token.
-        max_uses: Maximum number of uses (0 = unlimited).
-
-    Returns:
-        The created OrganizationToken.
-
-    Raises:
-        ValueError: If both grants_membership and grants_staff_status are False.
-        ValueError: If grants_membership is True and membership_tier is None.
-    """
-    if not grants_membership and not grants_staff_status:
-        raise ValueError("At least one of grants_membership or grants_staff_status must be True")
-    if grants_membership and membership_tier is None:
-        raise ValueError("membership_tier is required when grants_membership is True")
-    if isinstance(duration, int):
-        duration = timedelta(minutes=duration)
-    expires_at = timezone.now() + duration if duration else None
-    return OrganizationToken.objects.create(
-        name=name,
-        issuer=issuer,
-        organization=organization,
-        expires_at=expires_at,
-        max_uses=max_uses,
-        grants_membership=grants_membership,
-        grants_staff_status=grants_staff_status,
-        membership_tier=membership_tier,
-    )
-
-
-def _resolve_membership_tier(organization: Organization, tier_id: UUID | None) -> MembershipTier | None:
-    """Resolve a ``MembershipTier`` belonging to ``organization`` by id."""
-    if tier_id is None:
-        return None
-    return get_object_or_404(MembershipTier, pk=tier_id, organization=organization)
-
-
-def create_organization_token_from_payload(
-    *,
-    organization: Organization,
-    requested_by: RevelUser,
-    payload: schema.OrganizationTokenCreateSchema,
-) -> OrganizationToken:
-    """Create an organization token from an API payload.
-
-    Encapsulates the controller-side work that used to live in
-    ``create_organization_token`` view: owner-only privilege guard for
-    staff-granting tokens, membership-tier resolution, and dispatch to the
-    primitive ``create_organization_token`` constructor.
-
-    Args:
-        organization: The organization the token belongs to.
-        requested_by: The authenticated user issuing the create request.
-        payload: Validated create-schema from the API.
-
-    Returns:
-        The created ``OrganizationToken``.
-
-    Raises:
-        OrganizationTokenStaffGrantForbidden: When a non-owner tries to create
-            a staff-granting token.
-    """
-    if payload.grants_staff_status and organization.owner_id != requested_by.pk:
-        raise OrganizationTokenStaffGrantForbidden
-
-    data = payload.model_dump(exclude_unset=True)
-    tier_id = data.pop("membership_tier_id", None)
-    membership_tier = _resolve_membership_tier(organization, tier_id)
-
-    return create_organization_token(
-        organization=organization,
-        issuer=requested_by,
-        membership_tier=membership_tier,
-        **data,
-    )
-
-
-@transaction.atomic
-def update_organization_token(
-    token: OrganizationToken,
-    *,
-    requested_by: RevelUser,
-    payload: schema.OrganizationTokenUpdateSchema,
-) -> OrganizationToken:
-    """Update an organization token from an API payload.
-
-    Enforces the cross-field invariant (``grants_membership`` or
-    ``grants_staff_status`` must remain ``True``) and the owner-only guard
-    for any update that touches a staff-granting token (either the existing
-    token already grants staff, or the update would grant staff). Resolves
-    ``membership_tier_id`` to the corresponding ``MembershipTier`` and
-    persists changed fields only.
-
-    Args:
-        token: The token being updated.
-        requested_by: The authenticated user issuing the update request.
-        payload: Validated update-schema from the API.
-
-    Returns:
-        The refreshed ``OrganizationToken``.
-
-    Raises:
-        OrganizationTokenGrantInvariantError: When the resulting state would
-            disable both ``grants_membership`` and ``grants_staff_status``.
-        OrganizationTokenMembershipTierRequiredError: When the resulting state
-            would have ``grants_membership=True`` but no ``membership_tier``.
-        OrganizationTokenStaffGrantForbidden: When the requester is not the
-            organization owner but the token grants (or would grant) staff
-            status.
-    """
-    data = payload.model_dump(exclude_unset=True)
-
-    resulting_grants_staff = data.get("grants_staff_status", token.grants_staff_status)
-    resulting_grants_membership = data.get("grants_membership", token.grants_membership)
-    if not resulting_grants_membership and not resulting_grants_staff:
-        raise OrganizationTokenGrantInvariantError
-
-    touches_staff_grant = token.grants_staff_status or resulting_grants_staff
-    if touches_staff_grant and token.organization.owner_id != requested_by.pk:
-        raise OrganizationTokenStaffGrantForbidden
-
-    if "membership_tier_id" in data:
-        tier_id = data.pop("membership_tier_id")
-        data["membership_tier"] = _resolve_membership_tier(token.organization, tier_id)
-
-    # Resulting tier state: explicit value in this update, else the token's current tier.
-    resulting_tier = data["membership_tier"] if "membership_tier" in data else token.membership_tier
-    if resulting_grants_membership and resulting_tier is None:
-        raise OrganizationTokenMembershipTierRequiredError
-
-    if not data:
-        return token
-
-    for field, value in data.items():
-        setattr(token, field, value)
-    token.save(update_fields=list(data.keys()))
-    return token
-
-
-def delete_organization_token(
-    token: OrganizationToken,
-    *,
-    requested_by: RevelUser,
-) -> None:
-    """Delete an organization token, enforcing the owner-only guard.
-
-    Args:
-        token: The token to delete.
-        requested_by: The authenticated user issuing the delete request.
-
-    Raises:
-        OrganizationTokenStaffGrantForbidden: When the requester is not the
-            organization owner but the token grants staff status.
-    """
-    if token.grants_staff_status and token.organization.owner_id != requested_by.pk:
-        raise OrganizationTokenStaffGrantForbidden
-    token.delete()
-
-
-class OrgTokenRejection(t.NamedTuple):
-    """Why an organization token was rejected, plus the org it belongs to."""
-
-    reason: t.Literal["expired", "used_up"]
-    organization_id: UUID
-
-
-def get_organization_token(token: str) -> OrganizationToken | None:
-    """Retrieve an OrganizationToken by its ID.
-
-    Returns the token only if it is still valid: not expired and not
-    exhausted (``uses < max_uses``, or ``max_uses == 0`` for unlimited).
-    """
-    return (
-        OrganizationToken.objects.select_related("organization")
-        .filter(
-            Q(expires_at__isnull=True) | Q(expires_at__gt=timezone.now()),
-            Q(max_uses=0) | Q(uses__lt=F("max_uses")),
-            pk=token,
-        )
-        .first()
-    )
-
-
-def get_org_token_rejection_reason(token: str) -> OrgTokenRejection | None:
-    """Diagnose why an organization token is no longer valid.
-
-    Called only after get_organization_token() returned None to distinguish
-    "token doesn't exist" from "token expired / used up".
-    """
-    org_token = (
-        OrganizationToken.objects.only("expires_at", "uses", "max_uses", "organization_id").filter(pk=token).first()
-    )
-    if org_token is None:
-        return None
-    if org_token.expires_at and org_token.expires_at <= timezone.now():
-        return OrgTokenRejection(reason="expired", organization_id=org_token.organization_id)
-    if org_token.max_uses and org_token.uses >= org_token.max_uses:
-        return OrgTokenRejection(reason="used_up", organization_id=org_token.organization_id)
-    return None
-
-
-@transaction.atomic
-def claim_invitation(user: RevelUser, token: str) -> Organization | None:
-    """Claim an invitation given a Token."""
-    organization_token = (
-        OrganizationToken.objects.select_for_update()
-        .select_related("organization")
-        .filter(Q(expires_at__isnull=True) | Q(expires_at__gt=timezone.now()), pk=token)
-        .first()
-    )
-    if organization_token is None:
-        return None
-    if organization_token.max_uses and organization_token.uses >= organization_token.max_uses:
-        return None
-
-    # A hard-blacklisted (banned) user must not be able to claim their way back in.
-    # The membership path is blocked implicitly by colliding with the BANNED member row,
-    # but OrganizationStaff has no status field, so a banned ex-staffer could otherwise
-    # re-claim a grants_staff_status token and regain staff (bypassing BlacklistGate).
-    # Guard both grant paths explicitly, mirroring request_membership.
-    if blacklist_service.check_user_hard_blacklisted(user, organization_token.organization):
-        return None
-
-    if organization_token.grants_staff_status:
-        _, created = OrganizationStaff.objects.get_or_create(organization=organization_token.organization, user=user)
-    elif organization_token.grants_membership:
-        # Create member with tier if specified
-        defaults = {}
-        if organization_token.membership_tier:
-            defaults["tier"] = organization_token.membership_tier
-
-        _, created = OrganizationMember.objects.get_or_create(
-            organization=organization_token.organization, user=user, defaults=defaults
-        )
-    else:
-        return None
-
-    if not created:
-        return None
-    OrganizationToken.objects.filter(pk=token).update(uses=F("uses") + 1)
-    return organization_token.organization
-
-
-def add_member(organization: Organization, user: RevelUser, tier: MembershipTier) -> OrganizationMember:
-    """Add a member to an organization.
-
-    Args:
-        organization: The organization to add the member to.
-        user: The user to add as a member.
-        tier: The membership tier to assign to the member.
-
-    Returns:
-        The created OrganizationMember instance.
-
-    Raises:
-        AlreadyMemberError: If the user is already a member of the organization.
-    """
-    if OrganizationMember.objects.filter(organization=organization, user=user).exists():
-        raise AlreadyMemberError(str(_("User is already a member of this organization.")))
-    return OrganizationMember.objects.create(organization=organization, user=user, tier=tier)
-
-
-def remove_member(organization: Organization, user: RevelUser) -> None:
-    """Remove a member from an organization."""
-    member = get_object_or_404(OrganizationMember, organization=organization, user=user)
-    member.delete()
-
-
-def update_member(
-    member: OrganizationMember,
-    *,
-    status: OrganizationMember.MembershipStatus | None = None,
-    tier: MembershipTier | None = None,
-    clear_tier: bool = False,
-) -> OrganizationMember:
-    """Update a member's status and/or tier.
-
-    Args:
-        member: The OrganizationMember instance to update
-        status: New membership status (if provided)
-        tier: New membership tier (if provided)
-        clear_tier: If True, sets tier to None
-
-    Returns:
-        Updated OrganizationMember instance
-    """
-    updated_fields = []
-
-    if status is not None:
-        member.status = status
-        updated_fields.append("status")
-
-    if clear_tier:
-        member.tier = None
-        updated_fields.append("tier")
-    elif tier is not None:
-        member.tier = tier
-        updated_fields.append("tier")
-
-    if updated_fields:
-        member.save(update_fields=updated_fields)
-
-    return member
-
-
-def add_staff(
-    organization: Organization, user: RevelUser, permissions: PermissionsSchema | None = None
-) -> OrganizationStaff:
-    """Add a staff member to an organization."""
-    if OrganizationStaff.objects.filter(organization=organization, user=user).exists():
-        raise AlreadyMemberError(str(_("User is already a staff member of this organization.")))
-
-    permission_data = permissions.model_dump(mode="json") if permissions else _get_default_permissions()
-
-    return OrganizationStaff.objects.create(organization=organization, user=user, permissions=permission_data)
-
-
-def remove_staff(organization: Organization, user: RevelUser) -> None:
-    """Remove a staff member from an organization."""
-    staff = get_object_or_404(OrganizationStaff, organization=organization, user=user)
-    staff.delete()
-
-
-def update_staff_permissions(staff_member: OrganizationStaff, permissions: PermissionsSchema) -> OrganizationStaff:
-    """Update the permissions for a staff member."""
-    staff_member.permissions = permissions.model_dump(mode="json")
-    staff_member.save()
-    return staff_member
-
-
-@transaction.atomic
-def update_contact_email(organization: Organization, new_email: str, requester: RevelUser) -> str:
-    """Update organization contact email and send verification.
-
-    Self-heal: changing to a new (unverified) email forces ``contact_method`` back
-    to ``NONE``. This preserves the resolver invariant that a non-NONE contact
-    method always points at a verified email, without surprising owners with a
-    hard 400 on a normal email-change flow.
-
-    Args:
-        organization: The organization to update
-        new_email: The new contact email address
-        requester: The user requesting the change
-
-    Returns:
-        The verification token to be sent via email
-
-    Raises:
-        HttpError: If the email is the same as current one
-    """
-    # Check if email is different from current
-    if organization.contact_email and organization.contact_email.lower() == new_email.lower():
-        raise HttpError(400, str(_("This is already the contact email for this organization.")))
-
-    # Check if new email matches requester's verified email
-    if new_email.lower() == requester.email.lower() and requester.email_verified:
-        # Automatically verify since it matches the user's verified email
-        organization.contact_email = new_email
-        organization.contact_email_verified = True
-        organization.save(update_fields=["contact_email", "contact_email_verified"])
-        return ""  # No token needed
-
-    # Update the email but mark as unverified, and reset contact_method.
-    organization.contact_email = new_email
-    organization.contact_email_verified = False
-    organization.contact_method = Organization.ContactMethod.NONE
-    organization.save(update_fields=["contact_email", "contact_email_verified", "contact_method"])
-
-    # Create verification token and send email
-    return _create_and_send_contact_email_verification(organization, new_email, requester)
-
-
-def create_contact_message(
-    organization: Organization,
-    sender: RevelUser,
-    *,
-    subject: str,
-    message: str,
-) -> OrganizationContactMessage:
-    """Persist a contact-form submission for an organization.
-
-    The caller is responsible for verifying that the org's ``contact_method``
-    is ``FORM``; this function does not re-check the precondition. Sender's
-    email is captured from the user record so that downstream delivery uses
-    the verified address regardless of later account changes.
-    """
-    if organization.contact_method != Organization.ContactMethod.FORM:
-        raise HttpError(400, str(_("Contact form is not enabled for this organization.")))
-
-    return OrganizationContactMessage.objects.create(
-        organization=organization,
-        sender=sender,
-        sender_email_snapshot=sender.email,
-        subject=subject,
-        message=message,
-    )
-
-
-def validate_contact_method(organization: Organization, contact_method: Organization.ContactMethod) -> None:
-    """Validate that the requested ``contact_method`` is allowed for this org.
-
-    Setting EMAIL or FORM requires a verified contact email. NONE is always
-    allowed. Raises ``HttpError(400)`` if the rule is violated.
-    """
-    if contact_method == Organization.ContactMethod.NONE:
-        return
-    if not (organization.contact_email and organization.contact_email_verified):
-        raise HttpError(
-            400,
-            str(_("A verified contact email is required to enable a contact method other than NONE.")),
-        )
-
-
-@transaction.atomic
-def update_organization(
-    organization: Organization, payload: schema.OrganizationEditSchema, *, requester: RevelUser
-) -> Organization:
-    """Update an organization's editable fields.
-
-    Validates ``contact_method`` against the organization's verified-email state
-    and ``revenue_report_cadence`` against the presence of ``billing_email``
-    before persisting. Other fields are applied via ``update_db_instance``.
-
-    ``revenue_report_cadence`` controls scheduled financial-report delivery and is
-    owner-only, consistent with the owner-only org revenue endpoints. ``edit_organization``
-    is staff-grantable, so a non-owner attempting to change the cadence is rejected here
-    regardless of which permission got them to this endpoint.
-    """
-    from events.service import update_db_instance
-
-    data = payload.model_dump(exclude_unset=True)
-    if "contact_method" in data:
-        validate_contact_method(organization, Organization.ContactMethod(data["contact_method"]))
-
-    if (
-        "revenue_report_cadence" in data
-        and data["revenue_report_cadence"] != organization.revenue_report_cadence
-        and organization.owner_id != requester.pk
-    ):
-        raise RevenueReportCadenceOwnerOnlyError
-
-    # Apply cadence from payload (if present) before the billing_email check,
-    # so the check reflects the intended post-save state.
-    effective_cadence = data.get("revenue_report_cadence", organization.revenue_report_cadence)
-    if effective_cadence != Organization.RevenueReportCadence.NONE and not organization.billing_email:
-        raise ValidationError({"revenue_report_cadence": "A billing email is required to enable scheduled reports."})
-
-    return update_db_instance(organization, payload)
-
-
-@transaction.atomic
-def verify_contact_email(token: str) -> Organization:
-    """Verify an organization's contact email.
-
-    Args:
-        token: The verification token
-
-    Returns:
-        The organization with verified contact email
-
-    Raises:
-        HttpError: If token is invalid or organization not found
-    """
-    payload = token_to_payload(token, schema.VerifyOrganizationContactEmailJWTPayloadSchema)
-    check_blacklist(payload.jti)
-
-    organization = Organization.objects.filter(id=payload.organization_id).first()
-    if not organization:
-        raise HttpError(400, str(_("Organization not found.")))
-
-    # Check that the email in the token matches the current contact email
-    if not organization.contact_email or organization.contact_email.lower() != payload.email.lower():
-        raise HttpError(400, str(_("This verification link is for a different email address.")))
-
-    # Mark as verified
-    blacklist_token(token)
-    organization.contact_email_verified = True
-    organization.save(update_fields=["contact_email_verified"])
-
-    return organization
